- Introduce new parent (abstract) class StepPayload for Input / Output - Pull up the metadata (anygetter / anysetter) to the StepPayload to enable setting metadata for Inputs as well - Forward the metadata for "InputFromOutput" delegator - Rename old method Step#loadInputs(inputSet) to #loadOutputsFor(inputSet) and make it private - Add new method Step#loadInputs(inputSet) to load all inputs for a specified inputSet (to be used by subclasses)
